+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Format `<github issue/pr number>: <short description>`.
 * [#2759](https://github.com/kroxylicious/kroxylicious/pull/2759): Remove kroxylicious-sample and document how to use io.kroxylicious:kroxylicious-filter-archetype
 * [#2671](https://github.com/kroxylicious/kroxylicious/pull/2671): SASL inspection filter supoporting PLAIN, SCRAM-SHA-256 and SCRAM-SHA-512.
 * [#2681](https://github.com/kroxylicious/kroxylicious/pull/2681): Create a maven archetype for filter development io.kroxylicious:kroxylicious-filter-archetype
+* [#143](https://github.com/kroxylicious/kroxylicious/issues/143): Add support for Netty metrics 
 
 ## 0.16.0