+++ CHANGELOG.md
@@ -5,7 +5,7 @@ For changes that effect a public API, the [deprecation policy](./DEV_GUIDE.md#de
 
 Format `<github issue/pr number>: <short description>`.
 
-## SNAPSHOT
+## 0.17.0
 
 * [#2830](https://github.com/kroxylicious/kroxylicious/pull/2830): Refactor authentication configuration class names and optional fields in Azure KMS provider for Record Encryption
 * [#2784](https://github.com/kroxylicious/kroxylicious/issues/2784): Add support for Managed Identity authentication to Azure KMS provider for Record Encryption